move code into hex arch
