+++ rust/cloud-storage/Cargo.toml
@@ -20,6 +20,7 @@ members = [
   "connection_gateway",
   "connection_gateway_client",
   "connection_gateway_models",
+  "contacts",
   "contacts_db_client",
   "contacts_service",
   "convert_service",
