+++ rust/cloud-storage/contacts/src/domain.rs
@@ -1,0 +1,6 @@
+/// Domain models for contact graphs and users.
+pub mod models;
+/// Port traits defining boundaries between domain and infrastructure.
+pub mod ports;
+/// Domain service implementing business logic.
+pub mod service;
