+++ rust/cloud-storage/contacts/src/domain/models.rs
@@ -1,0 +1,4 @@
+/// Generic graph data structures (vertices and edges).
+pub mod graph;
+/// User-specific graph abstractions (groups, connections).
+pub mod user;
