+++ rust/cloud-storage/contacts/src/domain/models/user/test.rs
@@ -1,0 +1,131 @@
+use super::*;
+
+async fn unpack<'a, 'b>(
+    msg: &'a ConnectionsMessage,
+    users: &'b Vec<UserVertex>,
+) -> (Group, Vec<Connection<'b>>)
+where
+    'a: 'b,
+{
+    let mut group = Group::default();
+    for user in users {
+        group.participants.insert(user.clone());
+    }
+
+    let connections = unpack_connections(msg, users).await;
+
+    (group, connections)
+}
+
+fn generate_test_users() -> Vec<Vertex<User>> {
+    let ulist: Vec<Vertex<User>> = [
+        ("oceanus", "ff038d36-1aef-461a-8aa8-34001fa1abad"),
+        ("tethys", "5ab8c770-f2cb-4c6c-bc08-ae64569e324c"),
+        ("hyperion", "d44caada-98c0-49eb-ab20-6851b824983a"),
+        ("theia", "79a5557b-7827-4e2e-a6ae-f0935cdb762e"),
+        ("coeus", "c3f4d826-f8fd-478a-aa66-b5b6bb370cbc"),
+        ("phoebe", "c3b1970f-18ee-4dfa-b5fb-e8240e28e51d"),
+        ("cornus", "9effe035-bb12-4fcc-b479-800e1c2551a8"),
+        ("rhea", "ae2c090c-e478-4454-a001-3df458bf1fe4"),
+        ("mnemosyne", "b4e6267e-83c4-427d-88f4-40483f4b97e6"),
+        ("themis", "f5263664-b82f-41f5-bd4c-65e445f43e54"),
+        ("crius", "083a6148-26c6-4a59-9106-dbfb82579edc"),
+        ("iapetus", "6be3aef7-0701-4f0c-be6e-750f23ae953c"),
+    ]
+    .iter()
+    .map(|(_, uuid)| {
+        Vertex::new(User {
+            id: uuid.to_string(),
+        })
+    })
+    .collect();
+
+    ulist
+}
+
+#[test]
+fn test_group() {
+    let mut g = Group::default();
+    let users = generate_test_users();
+    let bob = Vertex::new(User {
+        id: "52D09596-7F05-4956-B64C-977AB9E334F9".to_string(),
+    });
+
+    let nusers = users.len();
+    // Populate group with some users
+    for user in users {
+        g.participants.insert(user);
+    }
+
+    // append a new user
+    let new_connections = g.append(&bob);
+
+    assert_eq!(new_connections.len(), nusers);
+}
+
+#[test]
+fn test_group_generate() {
+    let mut g = Group::default();
+    let users = generate_test_users();
+
+    let nusers = users.len();
+    let nconnections = nusers * (nusers - 1) / 2;
+
+    // Populate group with some users
+    for user in users {
+        g.participants.insert(user);
+    }
+
+    let connections = g.generate();
+
+    assert_eq!(connections.len(), nconnections);
+}
+
+#[test]
+fn test_connections_message() {
+    let mut g = Group::default();
+    let users = generate_test_users();
+
+    let nusers = users.len();
+    let nconnections = nusers * (nusers - 1) / 2;
+
+    // Populate group with some users
+    for user in users {
+        g.participants.insert(user);
+    }
+    let connections = g.generate();
+    let msg = create_connections_message(&g, &connections);
+
+    assert_eq!(msg.users.len(), nusers);
+    assert_eq!(msg.connections.len(), nconnections);
+}
+
+#[tokio::test]
+async fn test_connections_message_unpack() {
+    let mut g = Group::default();
+    let users = generate_test_users();
+
+    let nusers = users.len();
+    let nconnections = nusers * (nusers - 1) / 2;
+
+    // Populate group with some users
+    for user in users {
+        g.participants.insert(user);
+    }
+    let connections = g.generate();
+    let msg = create_connections_message(&g, &connections);
+
+    let new_users = unpack_users(&msg).await;
+    let (new_group, new_connections) = unpack(&msg, &new_users).await;
+
+    // Make sure all participants are accounted for
+    for user in &g.participants {
+        assert!(
+            new_group.participants.contains(user),
+            "Could not find user {}",
+            user.data.id
+        );
+    }
+
+    assert_eq!(new_connections.len(), nconnections);
+}
