+++ rust/cloud-storage/contacts/src/outbound.rs
@@ -1,0 +1,4 @@
+/// Database-backed contacts repository.
+pub mod repository;
+/// Connection gateway notifier adapter.
+pub mod gateway;
