+++ rust/cloud-storage/contacts/tests/fixtures/connections_message.json
