+++ rust/cloud-storage/contacts/tests/fixtures/create_group.json
