+++ rust/cloud-storage/contacts_service/Cargo.toml
@@ -15,25 +15,18 @@ path = "src/openapi.rs"
 
 [dependencies]
 anyhow = { workspace = true }
-async-trait = { workspace = true }
 aws-sdk-secretsmanager = { workspace = true }
 aws-sdk-sqs = { workspace = true }
-axum = { workspace = true }
-axum-extra = { workspace = true }
 connection_gateway_client = { path = "../connection_gateway_client" }
-contacts_db_client = { path = "../contacts_db_client" }
-http-body-util = { workspace = true }
+contacts = { path = "../contacts", features = ["axum"] }
 macro_auth = { path = "../macro_auth" }
 macro_aws_config = { path = "../macro_aws_config" }
-macro_cors = { path = "../macro_cors" }
 macro_entrypoint = { path = "../macro_entrypoint" }
+model = { path = "../model" }
 macro_env = { path = "../macro_env" }
 macro_middleware = { path = "../macro_middleware", default-features = false, features = [
   "auth",
 ] }
-macro_user_id = { path = "../macro_user_id" }
-model = { path = "../model" }
-model_user = { path = "../model_user", features = ["axum"] }
 rate_limit = { path = "../rate_limit", features = ["axum"] }
 redis = { workspace = true }
 secretsmanager_client = { path = "../secretsmanager_client" }
@@ -42,15 +35,8 @@ serde_json = { workspace = true }
 sqlx = { workspace = true }
 sqs_worker = { path = "../sqs_worker" }
 tokio = { workspace = true }
-tower = { workspace = true }
 tracing = { workspace = true }
 utoipa = { workspace = true }
-utoipa-swagger-ui = { workspace = true }
 
 [features]
 mock = []
-
-[dev-dependencies]
-macro_auth = { path = "../macro_auth", features = ["testing"] }
-rate_limit = { path = "../rate_limit", features = ["axum", "testing"] }
-rootcause = { workspace = true }
