--- rust/cloud-storage/contacts_service/src/fixtures/user_list.sql
@@ -1,25 +1,0 @@
-insert into connections (user1, user2) values
--- user 51's connections
-(
-	'1af59c26-c5b3-480c-9ad1-b87c2a69e72c',
-	'51028bda-67f0-44df-aa21-5853963524f1'
-),
-(
-	'51028bda-67f0-44df-aa21-5853963524f1',
-	'654df835-bb97-48e0-9b5b-0ad71a608dbd'
-),
-(
-	'51028bda-67f0-44df-aa21-5853963524f1',
-	'80c0effd-1b9d-4eba-ad86-bc52c4a63294'
-),
-
--- unrelated connections to 51
-(
-	'80c0effd-1b9d-4eba-ad86-bc52c4a63294',
-	'99c2c451-b5a2-4f47-9a38-1e7e56e0474b'
-),
-(
-	'80c0effd-1b9d-4eba-ad86-bc52c4a63294',
-	'ce2759ca-2610-448b-b496-b82ae0b14e41'
-);
-
