--- rust/cloud-storage/contacts_service/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod graph;
-pub mod queue;
-pub mod user;
