+++ rust/cloud-storage/contacts_service/src/main.rs
@@ -1,28 +1,23 @@
 #![recursion_limit = "256"]
-mod api;
 mod config;
-mod graph;
-mod queue;
-mod user;
 use std::sync::Arc;
 
 use anyhow::Context;
 use config::{Config, Environment};
 use connection_gateway_client::client::ConnectionGatewayClient;
-use contacts_service::queue::MessageQueue;
-use macro_auth::middleware::decode_jwt::JwtValidationArgs;
+use contacts::domain::service::ContactsDomainService;
+use contacts::inbound::http::AppState;
+use contacts::inbound::worker::ContactsWorker;
+use contacts::outbound::gateway::ConnectionGatewayNotifier;
+use contacts::outbound::repository::DbContactsRepository;
 use macro_entrypoint::MacroEntrypoint;
 use macro_middleware::auth::internal_access::InternalApiSecretKey;
 use rate_limit::{RateLimitServiceImpl, RedisRateLimitAdapter};
 use secretsmanager_client::SecretManager;
-use sqlx::PgPool;
 use sqlx::postgres::PgPoolOptions;
 use sqs_worker::SQSWorker;
 
-use crate::api::Service;
-use crate::api::context::AppState;
-
-async fn connect_to_database(config: &Config) -> anyhow::Result<PgPool> {
+async fn connect_to_database(config: &Config) -> anyhow::Result<sqlx::PgPool> {
     let (min_connections, max_connections): (u32, u32) = match config.environment {
         Environment::Production => (5, 30),
         Environment::Develop => (1, 25),
@@ -60,7 +55,6 @@ async fn main() -> anyhow::Result<()> {
     let config = Config::from_env().context("expected to be able to generate config")?;
 
     let db = connect_to_database(&config).await?;
-    let db_clone = db.clone();
     let sqs_worker = create_sqs_worker(&config).await;
 
     let secretsmanager_client = secretsmanager_client::SecretsManager::new(
@@ -71,19 +65,30 @@ async fn main() -> anyhow::Result<()> {
         .get_maybe_secret_value(config.environment, InternalApiSecretKey::new()?)
         .await?;
 
-    let connection_gateway_client = config.connection_gateway_url.as_ref().map(|url| {
-        ConnectionGatewayClient::new(internal_api_secret.as_ref().to_string(), url.clone())
+    let notifier = config.connection_gateway_url.as_ref().map(|url| {
+        ConnectionGatewayNotifier::new(ConnectionGatewayClient::new(
+            internal_api_secret.as_ref().to_string(),
+            url.clone(),
+        ))
     });
 
-    let mut worker = MessageQueue::new(sqs_worker, db_clone, connection_gateway_client);
+    let repository = DbContactsRepository::new(db.clone());
+    let service = Arc::new(ContactsDomainService {
+        repository,
+        notifier,
+    });
 
+    let worker = ContactsWorker::new(sqs_worker, service.clone());
     tokio::spawn(async move {
         worker.poll().await;
     });
 
     let jwt_args =
-        JwtValidationArgs::new_with_secret_manager(config.environment, &secretsmanager_client)
-            .await?;
+        macro_auth::middleware::decode_jwt::JwtValidationArgs::new_with_secret_manager(
+            config.environment,
+            &secretsmanager_client,
+        )
+        .await?;
 
     let redis_client =
         redis::Client::open(config.redis_uri.as_str()).context("failed to create redis client")?;
@@ -94,12 +99,10 @@ async fn main() -> anyhow::Result<()> {
         },
     };
 
-    api::setup_and_serve(AppState {
-        config: Arc::new(config),
-        db: db.clone(),
+    contacts::inbound::http::setup_and_serve(AppState {
+        port: config.port,
         jwt_args,
-        internal_api_secret,
-        contacts_service: Arc::new(Service(db)),
+        contacts_service: service,
         rate_limit_service,
     })
     .await?;
