+++ rust/cloud-storage/contacts_service/src/openapi.rs
@@ -1,16 +1,9 @@
-#![allow(unused)]
-
-mod api;
-mod config;
-mod graph;
-mod queue;
-mod user;
-
+use contacts::inbound::http::ApiDoc;
 use utoipa::OpenApi;
 
 fn main() {
     println!(
         "{}",
-        api::swagger::ApiDoc::openapi().to_pretty_json().unwrap()
+        ApiDoc::openapi().to_pretty_json().unwrap()
     );
 }
